clean up code

+++ Leanplum-SDK/Classes/Features/Variables/LPVarCache.h
@@ -110,7 +110,7 @@ NS_SWIFT_NAME(define(name:value:kind:));
 
 - (LPSecuredVars *)securedVars;
 
-- (int)getActionDefinitionType:(NSString *)actionName;
+- (NSUInteger)getActionDefinitionType:(NSString *)actionName;
 
 @end
